Integrate workflow in main window (#181)
* Integrate init-workflow component with main window

* Removed unneeded header and sidebars, fix date dropdowns, fix start and end dates

* Remove redundant workflow assets
