+++ bcfms/media/js/views/components/plugins/init-workflow.js
@@ -1,15 +1,145 @@
+import createVueApplication from 'arches/arches/app/media/js/utils/create-vue-application';
+import { createRouter, createWebHistory } from 'vue-router';
+import BCFMSApp from '@/bcfms/App.vue';
+import { routes } from '@/bcfms/routes.ts';
+import { definePreset } from '@primeuix/themes';
+import Aura from '@primeuix/themes/aura';
+
 import ko from 'knockout';
-import arches from 'arches';
 import defaultInitWorkflowTemplate from 'templates/views/components/plugins/init-workflow.htm';
 
-const InitWorkflow = function (params) {
-    this.workflows = params.workflows.map(function (wf) {
-        wf.url = arches.urls.plugin(wf.slug);
-        return wf;
-    }, this);
-};
+const router = createRouter({
+    history: createWebHistory(),
+    routes,
+});
 
-export default ko.components.register('init-workflow', {
-    viewModel: InitWorkflow,
+const BCGovPreset = definePreset(Aura, {
+    semantic: {
+        primary: {
+            50: '{blue.50}',
+            100: '{blue.100}',
+            200: '{blue.200}',
+            300: '{blue.300}',
+            400: '{blue.400}',
+            500: '{blue.500}',
+            600: '{blue.600}',
+            700: '{blue.700}',
+            800: '{blue.800}',
+            900: '{blue.900}',
+            950: '{blue.950}',
+        },
+        colorScheme: {
+            light: {
+                color: '{gray.50}',
+                formField: {
+                    hoverBorderColor: '{primary.color}',
+                },
+            },
+            dark: {
+                formField: {
+                    hoverBorderColor: '{primary.color}',
+                },
+            },
+        },
+        list: {
+            option: {
+                padding: '0.2rem 0.75rem',
+            },
+        },
+    },
+    components: {
+        button: {
+            paddingX: '.75rem;',
+            paddingY: '0.1rem;',
+        },
+        card: {
+            titleFontSize: '1.0rem',
+        },
+        checkbox: {
+            root: {
+                width: '1.75rem',
+                height: '1.75rem',
+            },
+        },
+        radiobutton: {
+            root: {
+                sm: {
+                    width: '1.75rem',
+                    height: '1.75rem',
+                },
+            },
+        },
+        fieldset: {
+            colorScheme: {
+                light: {
+                    background: '{grey.50}',
+                    legendBackground: '{grey.50}',
+                },
+                dark: {
+                    background: '{grey.900}',
+                    legendBackground: '{grey.900}',
+                },
+            },
+            legendFontSize: '2.0rem',
+        },
+        inputtext: {
+            paddingX: '0.2rem',
+            paddingY: '0.2rem',
+        },
+        select: {
+            root: {
+                paddingX: '1.0rem',
+                paddingY: '0.5rem',
+            },
+            option: {
+                fontSize: '1.4rem',
+                paddingY: '0.2rem',
+                padding: '0.2rem 0.2rem',
+                list: {
+                    padding: '0.2rem 0.2rem',
+                },
+            },
+        },
+        panel: {
+            contentPadding: '1.0rem',
+            colorScheme: {
+                light: {
+                    background: '{grey.50}',
+                },
+                dark: {
+                    background: '#222',
+                },
+            },
+        },
+        stepper: {
+            stepNumber: {
+                size: '2.8rem',
+                fontSize: '1.8rem',
+            },
+            steppanel: {
+                background: '{grey.50}',
+            },
+        },
+    },
+});
+
+ko.components.register('init-workflow', {
+    viewModel: function () {
+        createVueApplication(BCFMSApp, {
+            theme: {
+                preset: BCGovPreset,
+                options: {
+                    darkModeSelector: 'system',
+                    cssLayer: {
+                        name: 'primevue',
+                        order: 'theme, base, primevue',
+                    },
+                },
+            },
+        }).then((vueApp) => {
+            vueApp.use(router);
+            vueApp.mount('#bcfms-mounting-point');
+        });
+    },
     template: defaultInitWorkflowTemplate,
 });
