+++ bcfms/src/bcfms/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { provide, ref } from 'vue';
-import { useRouter, useRoute } from 'vue-router';
+import { useRouter } from 'vue-router';
 import { useGettext } from 'vue3-gettext';
 import Toast from 'primevue/toast';
 import { useToast } from 'primevue/usetoast';
@@ -15,8 +15,6 @@ import {
 } from '@/bcgov_arches_common/constants.ts';
 import { routeNames } from '@/bcfms/routes.ts';
 import { fetchUser } from '@/bcgov_arches_common/api.ts';
-import PageHeader from '@/bcgov_arches_common/components/header/PageHeader.vue';
-import SideNav from '@/bcgov_arches_common/components/sidenav/SideNav.vue';
 import type { Ref } from 'vue';
 import type { Language, User } from '@/bcgov_arches_common/types.ts';
 
@@ -33,7 +31,6 @@ const systemLanguage = ENGLISH; // TODO: get from settings
 provide(systemLanguageKey, systemLanguage);
 
 const router = useRouter();
-const route = useRoute();
 const toast = useToast();
 const { $gettext } = useGettext();
 
@@ -81,23 +78,7 @@ router.beforeEach(async (to, _from, next) => {
 
 <template>
     <main>
-        <PageHeader
-            v-if="route.meta.shouldShowNavigation"
-            :route-names="routeNames"
-            system-name="BC Fossil Management System"
-        />
-        <div
-            style="
-                display: flex;
-                flex: auto;
-                margin-top: 50px;
-                flex-direction: row;
-            "
-        >
-            <SideNav
-                v-if="route.meta.shouldShowNavigation"
-                :route-names="routeNames"
-            />
+        <div style="display: flex; flex: auto; flex-direction: row">
             <div
                 class="bcgov-main-content"
                 style="flex: auto; background-color: #e9e9e9"
@@ -125,10 +106,6 @@ main {
 </style>
 
 <style>
-#bcfms-mounting-point {
-    font-size: 0.8rem;
-}
-
 .bcgov-vertical-steps > .p-steplist {
     flex-direction: column;
     align-items: flex-start;
@@ -146,11 +123,5 @@ main {
 .bcgov-stepper {
     display: flex;
     flex-direction: row;
-}
-
-.p-tooltip-text,
-.p-button-label,
-.p-inputtext {
-    font-size: 0.8rem !important;
 }
 </style>
