+++ bcfms/src/bcfms/ipa/pages/SubmitProject/SubmitProject.vue
@@ -350,10 +350,6 @@ li {
 }
 </style>
 <style>
-.p-select-label {
-    font-size: 0.8rem !important;
-}
-
 .language-selector {
     display: none !important;
 }
