+++ bcfms/templates/arches_urls.htm
@@ -9,7 +9,6 @@
 <div class="arches-urls"
      api_user="{% url 'api_user' %}"
      user_profile_manager="{% url 'user_profile_manager' %}"
-     dashboard="{% url 'submissions' %}"
      submit_ipa='{% url "submit_ipa" %}'
      submit_ipa_review='(pk) => { return "{%url "submit-ipa-review" "00000000-0000-0000-0000-000000000000" %}".replace("00000000-0000-0000-0000-000000000000", pk) }'
      ipas_for_review='{% url "ipas-for-review" %}'
