--- bcfms/templates/bcfms/root_vue_dev.htm
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-{% load django_vite %}
-<!--[if IE 8]>
-<html lang="en" class="ie8"> <![endif]-->
-<!--[if IE 9]>
-<html lang="en" class="ie9"> <![endif]-->
-<!--[if !IE]><!-->
-<html lang="en"> <!--<![endif]-->
-
-
-<head>
-
-    <title> Vue Dev Template </title>
-    <base href="http://localhost:5174/">
-
-    <!-- Meta -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="">
-    <meta name="author" content="">
-
-    <link rel="shortcut icon" href="/bc-fossil-management/static/favicons/favicon.ico" type="image/x-icon" />
-    <link rel="apple-touch-icon" href="/bc-fossil-management/static/favicons/apple-touch-icon.png" />
-    <link rel="apple-touch-icon" sizes="76x76" href="/bc-fossil-management/static/favicons/apple-touch-icon-76x76.png" />
-    <link rel="apple-touch-icon" sizes="120x120" href="/bc-fossil-management/static/favicons/apple-touch-icon-120x120.png" />
-    <link rel="apple-touch-icon" sizes="152x152" href="/bc-fossil-management/static/favicons/apple-touch-icon-152x152.png" />
-    <link rel="apple-touch-icon" sizes="180x180" href="/bc-fossil-management/static/favicons/apple-touch-icon-180x180.png" />
-
-
-    <link rel="stylesheet" href="/bc-fossil-management/static/node_modules/normalize.css/normalize.css" />
-    <link rel="stylesheet" href="/bc-fossil-management/static/node_modules/font-awesome/css/font-awesome.min.css" />
-    <link rel="stylesheet" href="/bc-fossil-management/static/node_modules/primeicons/primeicons.css" />
-
-    <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.2/css/bootstrap.min.css">
-    <link href="/bc-fossil-management/static/css/unify.css" rel="stylesheet">
-    <link href="/bc-fossil-management/static/css/index.css" rel="stylesheet">
-    <link href="/bc-fossil-management/static/node_modules/arches/arches/app/media/css/index.css" rel="stylesheet" />
-    <link href="/bc-fossil-management/static/css/project.css" rel="stylesheet">
-    <!--    <link href="/node_modules/bcgov_arches_common/bcgov_arches_common/media/css/bc_index.css" rel="stylesheet" />-->
-    <link href="/bc-fossil-management/static/css/accessibility.css" rel="stylesheet">
-    <style>
-        html {
-            font-size: 16px;
-        }
-    </style>
-
-</head>
-
-
-<body dir="ltr">
-
-    {% vite_hmr_client %}
-    {% vite_asset '/bc-fossil-management/static/build/js/views/root_vue.js'%}
-    <div id="bcfms-mounting-point" class="bcgov-vue-mounting-point"></div>
-
-    {% block arches_modules %}
-    {% include "arches_urls.htm" %}
-    {% endblock arches_modules %}
-
-    <div class="arches-urls" api_user="/bc-fossil-management/api/user/"
-        user_profile_manager="/bc-fossil-management/user" dashboard="/bc-fossil-management/submissions/">
-    </div>
-
-</body>
-
-</html>
